# perf/feature: new serialization format for constraint systems