--- a/constraint/bls12-377/system.go
+++ b/constraint/bls12-377/system.go
@@ -17,18 +17,17 @@
 package cs
 
 import (
-	"bytes"
 	"encoding/binary"
 	"errors"
 	"github.com/blang/semver/v4"
 	"github.com/fxamacker/cbor/v2"
-	"github.com/ronanh/intcomp"
 	"io"
 	"time"
 
 	"github.com/consensys/gnark/backend/witness"
 	"github.com/consensys/gnark/constraint"
 	csolver "github.com/consensys/gnark/constraint/solver"
+	"github.com/consensys/gnark/internal/backend/ioutils"
 	"github.com/consensys/gnark/logger"
 	"reflect"
 
@@ -153,20 +152,25 @@ func (cs *system) CurveID() ecc.ID {
 	return ecc.BLS12_377
 }
 
-func (cs *system) ToBytes() ([]byte, error) {
-	buf := make([]byte, 0, cs.GetNbInstructions()*32)
+// WriteTo encodes R1CS into provided io.Writer using cbor
+func (cs *system) WriteTo(w io.Writer) (int64, error) {
+	_w := ioutils.WriterCounter{W: w}
 
-	// write the gnark version Major, Minor and Patch
 	gnarkVersion := semver.MustParse(cs.GnarkVersion)
-	buf = binary.LittleEndian.AppendUint32(buf, uint32(gnarkVersion.Major))
-	buf = binary.LittleEndian.AppendUint32(buf, uint32(gnarkVersion.Minor))
-	buf = binary.LittleEndian.AppendUint32(buf, uint32(gnarkVersion.Patch))
+	if err := binary.Write(&_w, binary.LittleEndian, uint32(gnarkVersion.Major)); err != nil {
+		return _w.N, err
+	}
+	if err := binary.Write(&_w, binary.LittleEndian, uint32(gnarkVersion.Minor)); err != nil {
+		return _w.N, err
+	}
+	if err := binary.Write(&_w, binary.LittleEndian, uint32(gnarkVersion.Patch)); err != nil {
+		return _w.N, err
+	}
 
 	// write the binary-friendly part of the system
 	// * the packed instructions
 	// * the calldata
 	// * the levels
-	// * the coefficients
 
 	// prepare the []uint32 separated slices for the packed instructions
 	sBlueprintID := make([]uint32, len(cs.Instructions))
@@ -184,84 +188,122 @@ func (cs *system) ToBytes() ([]byte, error) {
 
 	// they compress very well due to their nature (sequential integers)
 	var buf32 []uint32
-	buf32 = intcomp.CompressUint32(sBlueprintID, buf32)
-	buf = appendUint32Slice(buf, buf32)
 
-	buf32 = buf32[:0]
-	buf32 = intcomp.CompressUint32(sConstraintOffset, buf32)
-	buf = appendUint32Slice(buf, buf32)
+	var err error
+	// write the levels
+	binary.Write(&_w, binary.LittleEndian, uint64(len(cs.Levels)))
+	for _, l := range cs.Levels {
+		buf32, err = ioutils.CompressAndWriteUints32(&_w, l, buf32)
+		if err != nil {
+			return _w.N, err
+		}
+	}
 
-	buf32 = buf32[:0]
-	buf32 = intcomp.CompressUint32(sWireOffset, buf32)
-	buf = appendUint32Slice(buf, buf32)
+	buf32, err = ioutils.CompressAndWriteUints32(&_w, sBlueprintID, buf32)
+	if err != nil {
+		return _w.N, err
+	}
+	buf32, err = ioutils.CompressAndWriteUints32(&_w, sConstraintOffset, buf32)
+	if err != nil {
+		return _w.N, err
+	}
+	buf32, err = ioutils.CompressAndWriteUints32(&_w, sWireOffset, buf32)
+	if err != nil {
+		return _w.N, err
+	}
 
-	buf64 := intcomp.CompressUint64(sStartCallData, nil)
-	buf = appendUint64Slice(buf, buf64)
+	err = ioutils.CompressAndWriteUints64(&_w, sStartCallData)
+	if err != nil {
+		return _w.N, err
+	}
 
 	// write the calldata as var uints
-	buf = binary.LittleEndian.AppendUint64(buf, uint64(len(cs.CallData)))
+	binary.Write(&_w, binary.LittleEndian, uint64(len(cs.CallData)))
+	buf := make([]byte, 0, len(cs.CallData)*binary.MaxVarintLen64)
 	for _, v := range cs.CallData {
 		buf = binary.AppendUvarint(buf, uint64(v))
 	}
-
-	// TODO levels
-	// TODO coefficients
+	binary.Write(&_w, binary.LittleEndian, uint64(len(buf)))
+	_w.Write(buf)
 
 	// the rest is encoded with CBOR
 	ts := getTagSet()
 	enc, err := cbor.CoreDetEncOptions().EncModeWithTags(ts)
 	if err != nil {
-		return nil, err
+		return _w.N, err
 	}
-	var bbuf bytes.Buffer
-	encoder := enc.NewEncoder(&bbuf)
+	encoder := enc.NewEncoder(&_w)
 
 	// encode our object
 	err = encoder.Encode(cs)
-	if err != nil {
-		return nil, err
-	}
-
-	// append the CBOR encoded part to the buffer
-	buf = append(buf, bbuf.Bytes()...)
-
-	return buf, nil
+	return _w.N, err
 }
 
-func (cs *system) FromBytes(buf []byte) error {
-	// read the gnark version Major, Minor and Patch
-	if len(buf) < 12 {
-		return errors.New("invalid buffer size")
+// ReadFrom attempts to decode R1CS from io.Reader using cbor
+func (cs *system) ReadFrom(r io.Reader) (int64, error) {
+	var major, minor, patch uint32
+	if err := binary.Read(r, binary.LittleEndian, &major); err != nil {
+		return 0, err
 	}
-	_ = binary.LittleEndian.Uint32(buf[:4])
-	_ = binary.LittleEndian.Uint32(buf[4:8])
-	_ = binary.LittleEndian.Uint32(buf[8:12])
+	if err := binary.Read(r, binary.LittleEndian, &minor); err != nil {
+		return 4, err
+	}
+	if err := binary.Read(r, binary.LittleEndian, &patch); err != nil {
+		return 8, err
+	}
 	// if major != 0 || minor <= 10 {
-	// 	return errors.New("unsupported gnark version")
+	// 	return 0, errors.New("unsupported gnark version")
 	// }
 
 	// read the binary-friendly part of the system
 	// * the packed instructions
 	// * the calldata
 	// * the levels
-	// * the coefficients
 
 	// prepare the []uint32 separated slices for the packed instructions
-	var sBlueprintID, sConstraintOffset, sWireOffset []uint32
-	var sStartCallData []uint64
+	read := int64(12)
+
+	// read the levels
+	var levelsLen uint64
+	if err := binary.Read(r, binary.LittleEndian, &levelsLen); err != nil {
+		return read, err
+	}
+	read += 8
+
+	cs.Levels = make([][]uint32, levelsLen)
+	for i := range cs.Levels {
+		n, l, err := ioutils.ReadAndDecompressUints32(r)
+		if err != nil {
+			return read, err
+		}
+		read += int64(n)
+		cs.Levels[i] = l
+	}
 
 	// read the packed instructions
-	buf = buf[12:]
+	n, sBlueprintID, err := ioutils.ReadAndDecompressUints32(r)
+	if err != nil {
+		return read, err
+	}
+	read += int64(n)
 
-	var n int
-	n, sBlueprintID = readUint32Slice(buf)
-	buf = buf[n:]
-	n, sConstraintOffset = readUint32Slice(buf)
-	buf = buf[n:]
-	n, sWireOffset = readUint32Slice(buf)
-	buf = buf[n:]
-	n, sStartCallData = readUint64Slice(buf)
-	buf = buf[n:]
+	n, sConstraintOffset, err := ioutils.ReadAndDecompressUints32(r)
+	if err != nil {
+		return read, err
+	}
+	read += int64(n)
+
+	n, sWireOffset, err := ioutils.ReadAndDecompressUints32(r)
+	if err != nil {
+		return read, err
+	}
+	read += int64(n)
+
+	n, sStartCallData, err := ioutils.ReadAndDecompressUints64(r)
+	if err != nil {
+		return read, err
+	}
+	read += int64(n)
 
 	// rebuild the instructions
 	cs.Instructions = make([]constraint.PackedInstruction, len(sBlueprintID))
@@ -275,21 +317,31 @@ func (cs *system) FromBytes(buf []byte) error {
 	}
 
 	// read the calldata
-	calldataLen := binary.LittleEndian.Uint64(buf[:8])
-	buf = buf[8:]
-	cs.CallData = make([]uint32, calldataLen)
+	var calldataLen uint64
+	if err := binary.Read(r, binary.LittleEndian, &calldataLen); err != nil {
+		return read, err
+	}
+	read += 8
+	var buflen uint64
+	if err := binary.Read(r, binary.LittleEndian, &buflen); err != nil {
+		return read, err
+	}
+	read += 8
+	buf := make([]byte, buflen)
+	if _, err := io.ReadFull(r, buf); err != nil {
+		return read, err
+	}
 
+	cs.CallData = make([]uint32, calldataLen)
 	for i := range cs.CallData {
 		v, n := binary.Uvarint(buf)
 		if n <= 0 {
-			return errors.New("invalid uvarint")
+			return read, errors.New("invalid uvarint")
 		}
+		read += int64(n)
 		cs.CallData[i] = uint32(v)
 		buf = buf[n:]
 	}
-
-	// TODO levels
-	// TODO coefficients
 
 	// the rest is encoded with CBOR
 	ts := getTagSet()
@@ -299,19 +351,19 @@ func (cs *system) FromBytes(buf []byte) error {
 	}.DecModeWithTags(ts)
 
 	if err != nil {
-		return err
+		return read, err
 	}
-	decoder := dm.NewDecoder(bytes.NewReader(buf))
+	decoder := dm.NewDecoder(r)
 
 	// initialize coeff table
 	cs.CoeffTable = newCoeffTable(0)
 
 	if err := decoder.Decode(&cs); err != nil {
-		return err
+		return read + int64(decoder.NumBytesRead()), err
 	}
 
 	if err := cs.CheckSerializationHeader(); err != nil {
-		return err
+		return read + int64(decoder.NumBytesRead()), err
 	}
 
 	switch v := cs.CommitmentInfo.(type) {
@@ -321,63 +373,7 @@ func (cs *system) FromBytes(buf []byte) error {
 		cs.CommitmentInfo = *v
 	}
 
-	return nil
-}
-
-func appendUint32Slice(buf []byte, slice []uint32) []byte {
-	buf = binary.LittleEndian.AppendUint64(buf, uint64(len(slice)))
-	for _, v := range slice {
-		buf = binary.LittleEndian.AppendUint32(buf, v)
-	}
-	return buf
-}
-
-func readUint32Slice(data []byte) (int, []uint32) {
-	length := binary.LittleEndian.Uint64(data[:8])
-	in := make([]uint32, length)
-	for i := range in {
-		in[i] = binary.LittleEndian.Uint32(data[8+i*4:])
-	}
-	nbBytesRead := 8 + len(in)*4
-	return nbBytesRead, intcomp.UncompressUint32(in, nil)
-}
-
-func appendUint64Slice(buf []byte, slice []uint64) []byte {
-	buf = binary.LittleEndian.AppendUint64(buf, uint64(len(slice)))
-	for _, v := range slice {
-		buf = binary.LittleEndian.AppendUint64(buf, v)
-	}
-	return buf
-}
-
-func readUint64Slice(data []byte) (int, []uint64) {
-	length := binary.LittleEndian.Uint64(data[:8])
-	in := make([]uint64, length)
-	for i := range in {
-		in[i] = binary.LittleEndian.Uint64(data[8+i*8:])
-	}
-	nbBytesRead := 8 + len(in)*8
-	return nbBytesRead, intcomp.UncompressUint64(in, nil)
-}
-
-// WriteTo encodes R1CS into provided io.Writer using cbor
-func (cs *system) WriteTo(w io.Writer) (int64, error) {
-	buf, err := cs.ToBytes()
-	if err != nil {
-		return 0, err
-	}
-	n, err := w.Write(buf)
-	return int64(n), err
-}
-
-// ReadFrom attempts to decode R1CS from io.Reader using cbor
-func (cs *system) ReadFrom(r io.Reader) (int64, error) {
-	buf, err := io.ReadAll(r)
-	if err != nil {
-		return 0, err
-	}
-	err = cs.FromBytes(buf)
-	return int64(len(buf)), err
+	return read + int64(decoder.NumBytesRead()), nil
 }
 
 func (cs *system) GetCoefficient(i int) (r constraint.Element) {