--- a/internal/backend/ioutils/testdata/fuzz/FuzzIntcomp32/040184a38a688233199662fb82e804832282bdcd652e915e40fd2b32d6b6d0f2
+++ b/internal/backend/ioutils/testdata/fuzz/FuzzIntcomp32/040184a38a688233199662fb82e804832282bdcd652e915e40fd2b32d6b6d0f2
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("go test fuzz v1\n[]byte(\"000\\xd7000 00XA000B0000\")\n")