--- a/internal/backend/ioutils/testdata/fuzz/FuzzIntcomp32/1264dffe19eef294b54757da360097ceac5138928762aec415ec72a20c66493b
+++ b/internal/backend/ioutils/testdata/fuzz/FuzzIntcomp32/1264dffe19eef294b54757da360097ceac5138928762aec415ec72a20c66493b
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("go test fuzz v1\n[]byte(\"0000\")\n")