# Description

This PR uses [integer compression techniques](https://github.com/ronanh/intcomp)  to serialize some constraint systems parts containing ids, offsets, etc and bypass our current cbor choice. (cbor is still used for the rest). 

As a bench, using `lzss.DecompressionTestCircuit` with params landing at 90M constraints (plonk), we go from 9225Mb to 2238Mb size, plus, we serialize (and most importantly deserialize) way faster:
```
benchmark                                      old ns/op        new ns/op      delta
BenchmarkSerialization/serialize_cbor-96       54962357901      3086254450     -94.38%
BenchmarkSerialization/serialize_cbor-96       41877637150      3117870148     -92.55%
BenchmarkSerialization/deserialize_cbor-96     100910397720     4267472009     -95.77%
BenchmarkSerialization/deserialize_cbor-96     117815108278     3776636214     -96.79%
```