@ivokub 2024-04-30
> Very nice optimizations! I think in general is good, but in a perfect scenario I would try to figure out if we can work with `io.Writer` and `io.Reader` instead of direct byte arrays as imo otherwise we have many copies of very similar structures in memory (constraint system itself, bytes when serializing them and also byte buffer for internal steps a la serializing calldata). But due to applying compression to uints we still cannot get away with all.
> 
> And I completely agree that it doesn't make sense for us to include additional compression in the implementation. It should be straightforward to add `gzip` etc. when the user is interested in it. Otherwise there is too many options (fast or good compression ...)

Yep, it makes sense that if we work with `[]byte` then can have parallel processing which is more important than saving a few GB in settings where memory usage is in tens of GBs anyway. Added the test corpus.

Good to merge from my side.